+++ src/components/UI/input.js
@@ -1,0 +1,13 @@
+import React from 'react'
+
+export default function Input(props) {
+    let placeholder = props.placeholder;
+  return (
+    <>
+        <input
+         type="text" 
+         className="w-full rounded-md py-2 bg-white placeholder:text-slate-400 placeholder:pl-2 text-slate-400 " 
+         placeholder={placeholder}/>
+    </>
+  )
+}
