+++ src/layouts/Dashboard/Dashboard.js
@@ -8,6 +8,5 @@ export default function Dashboard() {
     <Navigation>
          <Navbar />
     </Navigation>
-    
   )
 }
