+++ src/pages/_app.js
@@ -1,5 +1,12 @@
+import Dashboard from "@/layouts/Dashboard/Dashboard";
 import "@/styles/globals.css";
 
 export default function App({ Component, pageProps }) {
-  return <Component {...pageProps} />;
+  return (
+    <>
+    <Dashboard/> 
+    <Component {...pageProps} />;
+    </>
+
+  )
 }
